+++ src/interface/aimrt_module_c_interface/util/string.h
@@ -19,6 +19,18 @@ typedef struct {
 } aimrt_string_t;
 
 /**
+ * @brief String array
+ *
+ */
+typedef struct {
+  /// String array
+  aimrt_string_t* str_array;
+
+  /// Length of string array
+  size_t len;
+} aimrt_string_array_t;
+
+/**
  * @brief String view
  *
  */
@@ -29,6 +41,18 @@ typedef struct {
   /// Length of char buffer
   size_t len;
 } aimrt_string_view_t;
+
+/**
+ * @brief String view array
+ *
+ */
+typedef struct {
+  /// String view array
+  aimrt_string_view_t* str_array;
+
+  /// Length of string view array
+  size_t len;
+} aimrt_string_view_array_t;
 
 #ifdef __cplusplus
 }