+++ src/plugins/lcm_plugin/lcm_channel_backend.h
@@ -65,8 +65,7 @@ class LcmChannelBackend : public runtime::core::channel::ChannelBackendBase {
   std::string_view Name() const override { return "lcm"; }
 
   void Initialize(YAML::Node options_node,
-                  const runtime::core::channel::ChannelRegistry* channel_registry_ptr,
-                  runtime::core::channel::ContextManager* context_manager_ptr) override;
+                  const runtime::core::channel::ChannelRegistry* channel_registry_ptr) override;
   void Start() override;
   void Shutdown() override;
 
@@ -87,8 +86,6 @@ class LcmChannelBackend : public runtime::core::channel::ChannelBackendBase {
 
   Options options_;
   std::atomic<State> state_ = State::PreInit;
-
-  runtime::core::channel::ContextManager* context_manager_ptr_ = nullptr;
 
   executor::ExecutorRef sub_default_executor_ref_;                            // default executor
   std::function<executor::ExecutorRef(std::string_view)> get_executor_func_;  // can get executor by name