+++ src/plugins/net_plugin/http/http_channel_backend.h
@@ -39,8 +39,7 @@ class HttpChannelBackend : public runtime::core::channel::ChannelBackendBase {
   std::string_view Name() const override { return "http"; }
 
   void Initialize(YAML::Node options_node,
-                  const runtime::core::channel::ChannelRegistry* channel_registry_ptr,
-                  runtime::core::channel::ContextManager* context_manager_ptr) override;
+                  const runtime::core::channel::ChannelRegistry* channel_registry_ptr) override;
   void Start() override;
   void Shutdown() override;
 
@@ -61,7 +60,6 @@ class HttpChannelBackend : public runtime::core::channel::ChannelBackendBase {
   std::atomic<State> state_ = State::PreInit;
 
   const runtime::core::channel::ChannelRegistry* channel_registry_ptr_ = nullptr;
-  runtime::core::channel::ContextManager* context_manager_ptr_ = nullptr;
 
   std::shared_ptr<boost::asio::io_context> io_ptr_;
   std::shared_ptr<runtime::common::net::AsioHttpClientPool> http_cli_pool_ptr_;